requirements.txt
flask
gunicorn
flask-sqlalchemy
psycopg2-binary
python-dotenv